+++ CHANGELOG.md
@@ -115,6 +115,7 @@
   `with_step` allowing to change the range step.][3408]
 - [Aligned `Text.split` API with other methods and added `Text.lines`.][3415]
 - [Implemented a basic reader for the `Delimited` file format.][3424]
+- [Implemented a reader for the `Excel` file format.][3425]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -177,6 +178,7 @@
 [3408]: https://github.com/enso-org/enso/pull/3408
 [3415]: https://github.com/enso-org/enso/pull/3415
 [3424]: https://github.com/enso-org/enso/pull/3424
+[3425]: https://github.com/enso-org/enso/pull/3425
 
 #### Enso Compiler